graph refactored and tested
